requirements.txt
Pillow==5.0.0
qrcode==5.3
six==1.11.0
